+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [unreleased] - ReleaseDate
 
+## BREAKING CHANGES
+
+- features in Cargo.toml.
+
 ## Added
 
 - Added `AsBytes` for general purpose trait.